requirements.txt
requests==2.24.0
beautifulsoup4==4.9.2
imageio==2.9.0
lxml==4.5.2
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
selenium==3.141.0
